+++ src/collapse/props.ts
@@ -18,7 +18,7 @@ const props: TdCollapseProps = {
   /** 展开图标。值为 undefined 或 false 则不显示展开图标；值为 true 显示默认图标；值类型为函数，则表示完全自定义展开图标 */
   expandIcon: {
     type: Boolean,
-    value: true,
+    value: null,
   },
   /** 每个面板互斥展开，每次只展开一个面板 */
   expandMutex: {
@@ -28,11 +28,11 @@ const props: TdCollapseProps = {
   /** 展开的面板集合 */
   value: {
     type: Array,
+    value: null,
   },
   /** 展开的面板集合，非受控属性 */
   defaultValue: {
-    type: null,
-    value: undefined,
+    type: Array,
   },
 };