+++ src/collapse/type.ts
@@ -69,7 +69,6 @@ export interface TdCollapsePanelProps {
   };
   /**
    * 当前折叠面板展开图标，优先级大于 Collapse 的同名属性
-   * @default true
    */
   expandIcon?: {
     type: BooleanConstructor;
@@ -87,8 +86,7 @@ export interface TdCollapsePanelProps {
    */
   headerRightContent?: {
     type: StringConstructor;
-    optionalTypes: Array<BooleanConstructor>;
-    value?: string | boolean;
+    value?: string;
   };
   /**
    * 当前面板唯一标识，如果值为空则取当前面下标兜底作为唯一标识